Added error object for JSON query objects
