+++ include/skyr/percent_encoding/errors.hpp
@@ -12,19 +12,39 @@
 namespace skyr {
 inline namespace v1 {
 namespace percent_encoding {
-  /// Enumerates percent encoding errors
-  enum class percent_encode_errc {
-    /// Input was not a hex value
-    non_hex_input,
-    /// Overflow
-    overflow,
-    };
+/// Enumerates percent encoding errors
+enum class percent_encode_errc {
+  /// Input was not a hex value
+  non_hex_input,
+  /// Overflow
+  overflow,
+};
 
-  /// Creates a `std::error_code` given a `skyr::percent_encode_errc`
-  /// value
-  /// \param error A percent encoding error
-  /// \returns A `std::error_code` object
-  auto make_error_code(percent_encode_errc error) noexcept -> std::error_code;
+namespace details {
+class percent_encode_error_category : public std::error_category {
+ public:
+  [[nodiscard]] auto name() const noexcept -> const char * override {
+    return "percent encoding";
+  }
+
+  [[nodiscard]] auto message(int error) const noexcept -> std::string override {
+    switch (static_cast<percent_encode_errc>(error)) {
+      case percent_encode_errc::non_hex_input:return "Non hex input";
+      case percent_encode_errc::overflow:return "Overflow";
+      default:return "(Unknown error)";
+    }
+  }
+};
+}  // namespace
+
+/// Creates a `std::error_code` given a `skyr::percent_encode_errc`
+/// value
+/// \param error A percent encoding error
+/// \returns A `std::error_code` object
+inline auto make_error_code(percent_encode_errc error) noexcept -> std::error_code {
+  static const details::percent_encode_error_category category{};
+  return std::error_code(static_cast<int>(error), category);
+}
 }  // namespace percent_encoding
 }  // namespace v1
 }  // namespace skyr
