+++ src/filesystem/path.cpp
@@ -4,42 +4,18 @@
 // http://www.boost.org/LICENSE_1_0.txt)
 
 #include <skyr/filesystem/path.hpp>
-#include <skyr/percent_encoding/percent_decode_range.hpp>
+#include <skyr/percent_encoding/percent_decode.hpp>
 
 namespace skyr {
 inline namespace v1 {
 namespace filesystem {
-namespace {
-class path_error_category : public std::error_category {
- public:
-  [[nodiscard]] auto name() const noexcept -> const char * override {
-    return "url filesystem path";
-  }
-
-  [[nodiscard]] auto message(int error) const noexcept -> std::string override {
-    switch (static_cast<path_errc>(error)) {
-      case path_errc::invalid_path: return "Invalid path";
-      case path_errc::percent_decoding_error: return "Percent decoding error";
-      default: return "(Unknown error)";
-    }
-  }
-};
-
-const path_error_category category{};
-}  // namespace
-
-auto make_error_code(path_errc error) noexcept -> std::error_code {
-  return std::error_code(static_cast<int>(error), category);
-}
-
 auto from_path(const stdfs::path &path) -> tl::expected<url, std::error_code> {
   return make_url("file://" + path.generic_u8string());
 }
 
 auto to_path(const url &input) -> tl::expected<stdfs::path, std::error_code> {
   auto pathname = input.pathname();
-  auto decoded = skyr::percent_encoding::as<std::string>(
-      pathname | skyr::percent_encoding::view::decode);
+  auto decoded = skyr::percent_decode<std::string>(pathname);
   if (!decoded) {
     return tl::make_unexpected(
         make_error_code(path_errc::percent_decoding_error));
