+++ src/network/ipv6_address.cpp
@@ -18,38 +18,6 @@ inline namespace v1 {
 using namespace std::string_view_literals;
 
 namespace {
-class ipv6_address_error_category : public std::error_category {
- public:
-  [[nodiscard]] auto name() const noexcept -> const char * override {
-    return "ipv6 address";
-  }
-
-  [[nodiscard]] auto message(int error) const noexcept -> std::string override {
-    switch (static_cast<ipv6_address_errc>(error)) {
-      case ipv6_address_errc::does_not_start_with_double_colon:
-        return "IPv6 piece does not start with a double colon.";
-      case ipv6_address_errc::invalid_piece:
-        return "Invalid IPv6 piece.";
-      case ipv6_address_errc::compress_expected:
-        return "IPv6 address compression was expected.";
-      case ipv6_address_errc::empty_ipv4_segment:
-        return "IPv4 segment is empty.";
-      case ipv6_address_errc::invalid_ipv4_segment_number:
-        return "IPv4 segment number is invalid.";
-      default:
-        return "(Unknown error)";
-    }
-  }
-};
-
-const ipv6_address_error_category category{};
-}  // namespace
-
-auto make_error_code(ipv6_address_errc error) -> std::error_code {
-  return std::error_code(static_cast<int>(error), category);
-}
-
-namespace {
 inline auto hex_to_dec(char byte) noexcept {
   assert(std::isxdigit(byte, std::locale::classic()));
